Split appearance settings into theme/interface
